Crate lmdb_master_sys

Structs

MDB_cursor

MDB_env

MDB_envinfo

Information about the environment

MDB_stat

Statistics for a database in the environment

MDB_txn

MDB_val

Generic structure used for passing keys and data in and out of the database.

Constants

MDB_APPEND

MDB_APPENDDUP

MDB_BAD_DBI

MDB_BAD_RSLOT

MDB_BAD_TXN

MDB_BAD_VALSIZE

MDB_CORRUPTED

MDB_CP_COMPACT

MDB_CREATE

MDB_CURRENT

MDB_CURSOR_FULL

MDB_DBS_FULL

MDB_DUPFIXED

MDB_DUPSORT

MDB_FIRST

< Position at first key/data item

MDB_FIRST_DUP

< Position at first data item of current key. Only for #MDB_DUPSORT

MDB_FIXEDMAP

MDB_FMT_Z

MDB_GET_BOTH

< Position at key/data pair. Only for #MDB_DUPSORT

MDB_GET_BOTH_RANGE

< position at key, nearest data. Only for #MDB_DUPSORT

MDB_GET_CURRENT

< Return key/data at current cursor position

MDB_GET_MULTIPLE

< Return up to a page of duplicate data items from current cursor position. Move cursor to prepare for #MDB_NEXT_MULTIPLE. Only for #MDB_DUPFIXED

MDB_INCOMPATIBLE

MDB_INTEGERDUP

MDB_INTEGERKEY

MDB_INVALID

MDB_KEYEXIST

MDB_LAST

< Position at last key/data item

MDB_LAST_DUP

< Position at last data item of current key. Only for #MDB_DUPSORT

MDB_LAST_ERRCODE

MDB_MAPASYNC

MDB_MAP_FULL

MDB_MAP_RESIZED

MDB_MULTIPLE

MDB_NEXT

< Position at next data item

MDB_NEXT_DUP

< Position at next data item of current key. Only for #MDB_DUPSORT

MDB_NEXT_MULTIPLE

< Return up to a page of duplicate data items from next cursor position. Move cursor to prepare for #MDB_NEXT_MULTIPLE. Only for #MDB_DUPFIXED

MDB_NEXT_NODUP

< Position at first data item of next key

MDB_NODUPDATA

MDB_NOLOCK

MDB_NOMEMINIT

MDB_NOMETASYNC

MDB_NOOVERWRITE

MDB_NORDAHEAD

MDB_NOSUBDIR

MDB_NOSYNC

MDB_NOTFOUND

MDB_NOTLS

MDB_PAGE_FULL

MDB_PAGE_NOTFOUND

MDB_PANIC

MDB_PREV

< Position at previous data item

MDB_PREVSNAPSHOT

MDB_PREV_DUP

< Position at previous data item of current key. Only for #MDB_DUPSORT

MDB_PREV_MULTIPLE

< Position at previous page and return up to a page of duplicate data items. Only for #MDB_DUPFIXED

MDB_PREV_NODUP

< Position at last data item of previous key

MDB_PROBLEM

MDB_RDONLY

MDB_READERS_FULL

MDB_RESERVE

MDB_REVERSEDUP

MDB_REVERSEKEY

MDB_SET

< Position at specified key

MDB_SET_KEY

< Position at specified key, return key + data

MDB_SET_RANGE

< Position at first key greater than or equal to specified key.

MDB_SIZE_MAX

MDB_SUCCESS

MDB_TLS_FULL

MDB_TXN_FULL

MDB_VERSION_DATE

MDB_VERSION_MAJOR

MDB_VERSION_MINOR

MDB_VERSION_MISMATCH

MDB_VERSION_PATCH

MDB_WRITEMAP

Functions

mdb_cmp⚠

Compare two data items according to a particular database.

mdb_cursor_close⚠

Close a cursor handle.

mdb_cursor_count⚠

Return count of duplicates for current key.

mdb_cursor_dbi⚠

Return the cursor’s database handle.

mdb_cursor_del⚠

Delete current key/data pair

mdb_cursor_get⚠

Retrieve by cursor.

mdb_cursor_open⚠

Create a cursor handle.

mdb_cursor_put⚠

Store by cursor.

mdb_cursor_renew⚠

Renew a cursor handle.

mdb_cursor_txn⚠

Return the cursor’s transaction handle.

mdb_dbi_close⚠

Close a database handle. Normally unnecessary. Use with care:

mdb_dbi_flags⚠

Retrieve the DB flags for a database handle.

mdb_dbi_open⚠

Open a database in the environment.

mdb_dcmp⚠

Compare two data items according to a particular database.

mdb_del⚠

Delete items from a database.

mdb_drop⚠

Empty or delete+close a database.

mdb_env_close⚠

Close the environment and release the memory map.

mdb_env_copy⚠

Copy an LMDB environment to the specified path.

mdb_env_copy2⚠

Copy an LMDB environment to the specified path, with options.

mdb_env_copyfd⚠

Copy an LMDB environment to the specified file descriptor.

mdb_env_copyfd2⚠

Copy an LMDB environment to the specified file descriptor, with options.

mdb_env_create⚠

Create an LMDB environment handle.

mdb_env_get_fd⚠

Return the filedescriptor for the given environment.

mdb_env_get_flags⚠

Get environment flags.

mdb_env_get_maxkeysize⚠

Get the maximum size of keys and #MDB_DUPSORT data we can write.

mdb_env_get_maxreaders⚠

Get the maximum number of threads/reader slots for the environment.

mdb_env_get_path⚠

Return the path that was used in #mdb_env_open().

mdb_env_get_userctx⚠

Get the application information associated with the #MDB_env.

mdb_env_info⚠

Return information about the LMDB environment.

mdb_env_open⚠

Open an environment handle.

mdb_env_set_assert⚠

Set or reset the assert() callback of the environment. Disabled if liblmdb is built with NDEBUG.

mdb_env_set_flags⚠

Set environment flags.

mdb_env_set_mapsize⚠

Set the size of the memory map to use for this environment.

mdb_env_set_maxdbs⚠

Set the maximum number of named databases for the environment.

mdb_env_set_maxreaders⚠

Set the maximum number of threads/reader slots for the environment.

mdb_env_set_userctx⚠

Set application information associated with the #MDB_env.

mdb_env_stat⚠

Return statistics about the LMDB environment.

mdb_env_sync⚠

Flush the data buffers to disk.

mdb_get⚠

Get items from a database.

mdb_put⚠

Store items into a database.

mdb_reader_check⚠

Check for stale entries in the reader lock table.

mdb_reader_list⚠

Dump the entries in the reader lock table.

mdb_set_compare⚠

Set a custom key comparison function for a database.

mdb_set_dupsort⚠

Set a custom data comparison function for a #MDB_DUPSORT database.

mdb_set_relctx⚠

Set a context pointer for a #MDB_FIXEDMAP database’s relocation function.

mdb_set_relfunc⚠

Set a relocation function for a #MDB_FIXEDMAP database.

mdb_stat⚠

Retrieve statistics for a database.

mdb_strerror⚠

Return a string describing a given error code.

mdb_txn_abort⚠

Abandon all the operations of the transaction instead of saving them.

mdb_txn_begin⚠

Create a transaction for use with the environment.

mdb_txn_commit⚠

Commit all the operations of a transaction into the database.

mdb_txn_env⚠

Returns the transaction’s #MDB_env

mdb_txn_id⚠

Return the transaction’s ID.

mdb_txn_renew⚠

Renew a read-only transaction.

mdb_txn_reset⚠

Reset a read-only transaction.

mdb_version⚠

Return the LMDB library version information.

Type Aliases

MDB_assert_func

A callback function for most LMDB assert() failures, called before printing the message and aborting.

MDB_cmp_func

A callback function used to compare two keys in a database

MDB_cursor_op

Cursor Get operations.

MDB_dbi

A handle for an individual database in the DB environment.

MDB_msg_func

A callback function used to print a message from the library.

MDB_rel_func

A callback function used to relocate a position-dependent data item in a fixed-address database.

mdb_filehandle_t

mdb_mode_t

mdb_size_t

Unsigned type used for mapsize, entry counts and page/transaction IDs.