lmdb_master_sys

Function mdb_cursor_del

pub unsafe extern "C" fn mdb_cursor_del(
    cursor: *mut MDB_cursor,
    flags: c_uint,
) -> c_int

Delete current key/data pair

This function deletes the key/data pair to which the cursor refers. This does not invalidate the cursor, so operations such as MDB_NEXT can still be used on it. Both MDB_NEXT and MDB_GET_CURRENT will return the same record after this operation.

Arguments

cursor (direction in) - A cursor handle returned by #mdb_cursor_open()

flags (direction in) - Options for this operation. This parameter must be set to 0 or one of the values described here.

#MDB_NODUPDATA - delete all of the data items for the current key. This flag may only be specified if the database was opened with #MDB_DUPSORT.
# Returns
A non-zero error value on failure and 0 on success. Some possible errors are:

EACCES - an attempt was made to write in a read-only transaction.
EINVAL - an invalid parameter was specified.