lmdb_master_sys

Function mdb_env_copyfd

pub unsafe extern "C" fn mdb_env_copyfd(
    env: *mut MDB_env,
    fd: mdb_filehandle_t,
) -> c_int

Copy an LMDB environment to the specified file descriptor.

This function may be used to make a backup of an existing environment. No lockfile is created, since it gets recreated at need.

Note: This call can trigger significant file size growth if run in parallel with write transactions, because it employs a read-only transaction. See long-lived transactions under caveats_sec.

Arguments

env (direction in) - An environment handle returned by #mdb_env_create(). It must have already been opened successfully.

fd (direction in) - The filedescriptor to write the copy to. It must have already been opened for Write access.

Returns

A non-zero error value on failure and 0 on success.