lmdb_master_sys

Function mdb_env_create

pub unsafe extern "C" fn mdb_env_create(
    env: *mut *mut MDB_env,
) -> c_int

Create an LMDB environment handle.

This function allocates memory for a #MDB_env structure. To release the allocated memory and discard the handle, call #mdb_env_close(). Before the handle may be used, it must be opened using #mdb_env_open(). Various other options may also need to be set before opening the handle, e.g. #mdb_env_set_mapsize(), #mdb_env_set_maxreaders(), #mdb_env_set_maxdbs(), depending on usage requirements.

Arguments

env (direction out) - The address where the new handle will be stored

Returns

A non-zero error value on failure and 0 on success.